+++ src/test_fmindex-collection/suffixarray/checkCSA.cpp
@@ -41,7 +41,7 @@ TEST_CASE("check Compressed Suffix Array (CSA)", "[suffixarray][CSA]") {
     SECTION("sampling 3") {
         auto csa = fmc::CSA {sa, /*.samplingRate =*/ 3, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[2] = {0, 0};
         expected[3] = {1, 0};
         expected[6] = {1, 3};
@@ -52,7 +52,7 @@ TEST_CASE("check Compressed Suffix Array (CSA)", "[suffixarray][CSA]") {
     SECTION("sampling 4") {
         auto csa = fmc::CSA {sa, /*.samplingRate =*/ 4, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[2] = {0, 0};
         expected[3] = {1, 0};
         expected[4] = {1, 4};
@@ -63,7 +63,7 @@ TEST_CASE("check Compressed Suffix Array (CSA)", "[suffixarray][CSA]") {
     SECTION("sampling 5") {
         auto csa = fmc::CSA {sa, /*.samplingRate =*/ 5, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[0] = {1, 5};
         expected[1] = {0, 5};
         expected[2] = {0, 0};
@@ -74,7 +74,7 @@ TEST_CASE("check Compressed Suffix Array (CSA)", "[suffixarray][CSA]") {
     SECTION("sampling 8") {
         auto csa = fmc::CSA {sa, /*.samplingRate =*/ 8, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[2] = {0, 0};
         expected[3] = {1, 0};
         check(csa, expected);
@@ -115,7 +115,7 @@ TEST_CASE("check Dense Compressed Suffix Array (DenseCSA)", "[suffixarray][Dense
     SECTION("sampling 3") {
         auto csa = fmc::DenseCSA {sa, /*.samplingRate =*/ 3, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[2] = {0, 0};
         expected[3] = {1, 0};
         expected[6] = {1, 3};
@@ -126,7 +126,7 @@ TEST_CASE("check Dense Compressed Suffix Array (DenseCSA)", "[suffixarray][Dense
     SECTION("sampling 4") {
         auto csa = fmc::DenseCSA {sa, /*.samplingRate =*/ 4, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[2] = {0, 0};
         expected[3] = {1, 0};
         expected[4] = {1, 4};
@@ -137,7 +137,7 @@ TEST_CASE("check Dense Compressed Suffix Array (DenseCSA)", "[suffixarray][Dense
     SECTION("sampling 5") {
         auto csa = fmc::DenseCSA {sa, /*.samplingRate =*/ 5, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[0] = {1, 5};
         expected[1] = {0, 5};
         expected[2] = {0, 0};
@@ -148,7 +148,7 @@ TEST_CASE("check Dense Compressed Suffix Array (DenseCSA)", "[suffixarray][Dense
     SECTION("sampling 8") {
         auto csa = fmc::DenseCSA {sa, /*.samplingRate =*/ 8, inputSizes};
 
-        auto expected = std::vector<std::optional<std::tuple<size_t, size_t>>>(12);
+        auto expected = std::vector<std::optional<std::tuple<uint32_t, uint32_t>>>(12);
         expected[2] = {0, 0};
         expected[3] = {1, 0};
         check(csa, expected);
